+++ client/js/blockchain.js
@@ -14,9 +14,11 @@ export const processBlockchainInfo = async () => {
             slotSinceGenesis,
         } = blockchainInfo.data.bestChain[0].protocolState.consensusState
 
+        globalThis.epoch = epoch
+
         const total = (totalCurrency / 10**9)
 
-        $("#currency-total").text(total.format(0, null, " ", "."))
+        $("#total-supply").text(total.format(0, null, " ", "."))
         $("#epoch-number").text(epoch)
         $("#slot-number").text(slot)
         $("#slot-since-genesis").text(slotSinceGenesis)
@@ -30,5 +32,9 @@ export const processBlockSpeed = async () => {
 
     $("#block-speed").html(`<span class="text-bold fg-accent">${(blockSpeed / 60000).toFixed(2)}</span> minutes`)
 
-    setTimeout( () => processBlockSpeed(), blockSpeed ? blockSpeed : 300000 )
+    if (blockSpeed) {
+        globalThis.blockSpeed = blockSpeed
+    }
+
+    setTimeout( () => processBlockSpeed(), blockSpeed ? blockSpeed : 180000 )
 }