+++ server/node.mjs
@@ -41,8 +41,8 @@ query MyQuery {
 `;
 
 const queryBalance = `
-query MyQuery {
-  account(publicKey: "%PUBLIC_KEY%") {
+query ($publicKey: String!) {
+  account(publicKey: $publicKey) {
     balance {
       total
       blockHeight
@@ -103,9 +103,9 @@ query MyQuery {
 `;
 
 const queryBlockSpeed = `
-query MyQuery {
+query ($maxLength: Int) {
   version
-  bestChain(maxLength: %LENGTH%) {
+  bestChain(maxLength: $maxLength) {
     protocolState {
       blockchainState {
         date
@@ -118,7 +118,7 @@ query MyQuery {
 }
 `;
 
-async function fetchGraphQL(addr, query, operationName = "MyQuery", variables = {}) {
+async function fetchGraphQL(addr, query, variables = {}) {
     try {
         const result = await fetch(
             `http://${addr}/graphql`,
@@ -129,8 +129,7 @@ async function fetchGraphQL(addr, query, operationName = "MyQuery", variables = 
                 },
                 body: JSON.stringify({
                     query,
-                    operationName: operationName,
-                    variables: variables,
+                    variables,
                 })
             }
         )
@@ -143,12 +142,15 @@ async function fetchGraphQL(addr, query, operationName = "MyQuery", variables = 
 }
 
 async function getBlockSpeed(graphql, length){
-    let blocks = await fetchGraphQL(graphql, queryBlockSpeed.replace("%LENGTH%", length))
-    let chain, speed, begin, end
-    if (!blocks) {
+    let blocks = await fetchGraphQL(graphql, queryBlockSpeed, {maxLength: length})
+    if (!blocks && !blocks.data) {
         return 0
     }
-    chain = blocks.data.bestChain
+    const {bestChain: chain = []} = blocks.data
+    if (!chain.length) {
+        return 0
+    }
+    let speed, begin, end
     begin = +chain[0]["protocolState"]["blockchainState"]["date"]
     end = +chain[chain.length - 1]["protocolState"]["blockchainState"]["date"]
     speed = (end - begin) / length
@@ -156,11 +158,11 @@ async function getBlockSpeed(graphql, length){
 }
 
 export const nodeInfo = async (obj, config) => {
-    const {graphql, publicKey} = config
+    const {graphql, publicKey, publicKeyDelegators} = config
 
     switch (obj) {
         case 'node-status': return await fetchGraphQL(graphql, queryNodeStatus)
-        case 'balance': return publicKey ? await fetchGraphQL(graphql, queryBalance.replace("%PUBLIC_KEY%", publicKey)) : 0
+        case 'balance': return publicKey ? await fetchGraphQL(graphql, queryBalance, {publicKey}) : 0
         case 'blockchain': return await fetchGraphQL(graphql, queryBlockChain)
         case 'consensus': return await fetchGraphQL(graphql, queryConsensus)
         case 'block-speed': return await getBlockSpeed(graphql, 10)